+++ packages/data/src/entry-form/entry-form.component.ts
@@ -72,5 +72,6 @@ export class EntryFormComponent extends FormComponent {
       });
     });
     this.loaderService.wait(deletion, this.loader);
+    return deletion;
   }
 }